+++ app/src/main/java/org/simple/clinic/registration/confirmpin/RegistrationConfirmPinScreenKey.kt
@@ -2,15 +2,14 @@ package org.simple.clinic.registration.confirmpin
 
 import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
-import org.simple.clinic.R
-import org.simple.clinic.router.screen.FullScreenKey
+import org.simple.clinic.navigation.v2.ScreenKey
 import org.simple.clinic.user.OngoingRegistrationEntry
 
 @Parcelize
-data class RegistrationConfirmPinScreenKey(val registrationEntry: OngoingRegistrationEntry) : FullScreenKey {
+data class RegistrationConfirmPinScreenKey(val registrationEntry: OngoingRegistrationEntry) : ScreenKey() {
 
   @IgnoredOnParcel
   override val analyticsName = "Registration Confirm PIN"
 
-  override fun layoutRes() = R.layout.screen_registration_confirm_pin
+  override fun instantiateFragment() = RegistrationConfirmPinScreen()
 }